requirements.txt
torch>=1.12.0
torchaudio
openai-whisper
pyannote.audio
pydub
librosa
fastapi
uvicorn
pytest
